Set up the view classes
In order to display the kingdom information I have attached an observer
to the getKingdoms method in the view model and created the RecyclerView
adapter.
I have also upgraded the project to use Androidx for all libraries as
the project does not run otherwise.
The project is now ready to have the API added